middleware/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
langchain-groq
langchain-ollama
langchain-core
